Fixed AnalysisCompletionEvent callback
The `AnalysisCompletionEvent` class takes a callback as a parameter in its `__init__` method, but this callback is not added to the object as a bound method automatically in python 2 as it is in python 3. This means that the callback has no parameters passed to it, which prevents it from doing anything meaningful. This PR fixes this by binding the `callback` parameter to an initialized `AnalysisCompletionEvent`.