Implement .is_thunk in LLILFunction.
Allows a developer to use the LLILFunction or Function classes to
query whether or not a function is a thunk. I base this on whether
or not a Tailcall is observed within the first five expressions.

I have not seen an instance where a thunk appears greater than
this range. I did not want to check if the last instruction was a
Tailcall in the event that significant processing is performed
before the Tailcall, in which case, is it still a thunk?