Change type from array->string.
`enum` seems to be the list of strings that are possible values.
When the type is set to `array` and `enum` is present, Settings
fails to load. Additionally, I could not find any examples of
Settings where the type is set `array` and `enum` is added.

I confirmed that the list of strings defined in `enum` populate
a drop down list when the `type` is set to `string`.