Remove LLIL for LWL/LWR/SWL/SWR
This seemed to have assumed that if the code reaches the point where it's
lifting a lone LWL/LWR/SWL/SWR, then it didn't get caught by the code
that checked for pairs of LWL/LWR/SWL/SWR, and so it must be a 2-byte
read...however, this assumption doesn't hold for code like this:

```
swl v0, (v1)
<random other instructions>
swr v0, 3(v1)
```

In this case, the intent is still an unaligned store, just split up for
whatever reason the compiler did it, and this isn't a 2-byte store, which
is what the LLIL was.

The actual LLIL is going to be a lot trickier, but in the meantime it's
probably better to admit we can't lift this code, rather than lifting it
incorrectly.