Re-lift unaligned memory accesses
Note that LWL/LWR hasn't been tested as much, since the only places
these are used in the binary sample I have are right next to each other,
so they get treated as a pair and lifted together.

There didn't seem to be a clean way to lift these instructions that
wouldn't involve masking pointers, which isn't friendly to users or
automated analysis, so I settled on intrinsics.

One fidelity trade-off is that the LLIL indicates that the entire
32-bit/64-bit value is read/written to for each instruction.  The
decision to be optimistic about how many bytes get accessed was chosen to
help make sure structure xrefs get noticed.  For example, suppose there's
a 4-byte value at offset 7 of a struct; when this gets set by
SWL rX, 7(rY)/SWR rX, 0xa(rY), we want to make sure it gets caught by each
instruction, since each one writes that offset.

An unfortunate side effect is that in this example, the LLIL indicates
that offset 0xb also gets written to by the SWR instruction, which is not
true.  However, it's generally more annoying when xrefs aren't originally
found than having to filter through xrefs that do exist, so this
trade-off seemed worth it to me.