[SharedCache] Introduce a SymbolInfo type
Previously `SharedCache` used a mixture of `pair<uint64_t,
pair<BNSymbolType, std::string>>` and `Ref<Symbol>` to represent a
symbol. The former is hard to understand due to the nesting and lack of
names. The latter is expensive to create and its getters are expensive
due to the multiple allocations + `strlen` calls needed when converting
between core's string representation and `std::string`.

`SymbolInfo` wraps the same data the nested pairs contained into a
friendly struct with named fields. Beyond that it, it exposes an
`AsSymbol` function to construct a `Ref<Symbol>` where that is what
is required.

Code within `SharedCache` is updated to work with `SymbolInfo` until it
needs to call into a core API that requires a `Symbol`.

This change is only a small performance improvement at the moment. It's
more noticable once `SharedCache` no longer copies so much of its state
(#6393).