Remove m_reader / m_builder members from AnalysisContext
Constructing these objects are very expensive compared to the cost of
creating the `AnalysisContext`.

`m_reader` appears to be entirely unused so it is simply removed.

`m_builder` is only used within `Inform`. It may be appropriate to move
it to a local variable within that function, but I'm not sure how often
`Inform` is called. Given that the `Json::StreamWriterBuilder` is rather
expensive to construct I've instead opted to make it static so it is
initialized only on first use and then reused by later calls.