[SharedCache] Apply objc_msgSend call type overrides in more places
`fixObjCCallTypes` is updated to handle tail calls as well as regular
calls.

Additionally, if the selector address is not valid it now looks for the
`sel_` symbols that `DSCObjCProcessor::ReadMethodList` adds for
selectors whose names reside in regions that are not yet mapped. This
allows call type overrides to be applied even when a selector's name is
defined in a different image.