[SharedCache] Add a section for an image header
Similar to what is discussed in PR #6513 its useful to be able to see where an image header is as sometimes you want to inspect the load commands. Without a section its not very obvious where the header is.

The naming scheme of `${imageName}::__header` was chosen based on the naming choice of the DSC header.