Add bn::base::atomic_shared_ptr
`bn::base::atomic_shared_ptr<T>` is a type alias for
`std::atomic<std::shared_ptr<T>>` when the standard library supports it.
Otherwise, it is a class that holds a `std::shared_ptr<T>` and
implements the necessary interface in terms of the free atomic
operations.

This type is motivated by recent versions of libstdc++ having marked the
free atomic operations on `std::shared_ptr` as deprecated, resulting in
compiler warnings when they are used.